fix: resolve architecture gate failures
- Break NodeRenderer <-> DomNodeRenderer import cycle by passing ChildRenderer as prop
- Break forms index.ts <-> htmlContracts.ts cycle by moving prop schemas/types to forms/types.ts
- Fix no-css-var-fallbacks violation in DomNodeProperties.module.css
- Extract createDomNode/syncModuleOverlayHtmlFields to page-tree/domMutations.ts
  and ratchet mutations.ts module-size cap down to 766
- Grandfather agent executor.ts (734 lines) with note to extract DOM-native tool runners
- Fix Windows path bug in error-boundary-coverage test
- Sync plugin bootstrap artifacts via bootstrap:sync
- Use @core/page-tree barrel import in hybrid-dom-native-nodes architecture test
